[#36] refactor: 분리 - validation function
- 재사용을 위해 common.js로 validation 함수 분리
