+++ server/api/routes/record-reload.js
@@ -1,14 +1,12 @@
 const express = require('express');
 const router = express.Router();
 import { wrapAsync } from '../../util';
-
 import { isReloadInMonth } from '../../controller/record-reload';
-
-import { validateRecordParameter } from '../../validator/record';
+import { validateRecordReloadParameter } from '../../validator/record';
 
 router.get(
   '/:year/:month',
-  validateRecordParameter,
+  validateRecordReloadParameter,
   wrapAsync(isReloadInMonth),
 );
