+++ server/controller/record-reload.js
@@ -5,13 +5,12 @@ const isReloadInMonth = async function (req, res, next) {
   const month = req.params.month;
   const lastEditedAt = req.query.lastEditedAt;
   const lastUpdate = await getRecentUpdatedDate(year, month);
-  const isReload =
-    lastUpdate.length === 0 || !lastEditedAt
-      ? false
-      : Date.parse(lastUpdate) > Date.parse(lastEditedAt);
+  const isReload = Date.parse(lastUpdate) > Date.parse(lastEditedAt);
   res.send({
     success: true,
     isReload,
+    year,
+    month,
   });
 };
